+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please refer to the [official documentation](https://wilfred.readthedocs.io/en/l
 ### Bugs
 
 * Only perform Travis CI `pypi` deployment once, an error in the config caused the CI to deploy twice during the `v0.1.0` release.
+* [#10](https://github.com/wilfred-dev/wilfred/issues/10) Truncate custom startup commands, too long commands do no longer break table styling.
 
 ## v0.1.0 (released on 2020-01-11)