+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Please refer to the [official documentation](https://wilfred.readthedocs.io/en/l
 ### Added
 
 * [#14](https://github.com/wilfred-dev/wilfred/issues/14) Added ability to restart servers using `wilfred restart <name>`.
+* [#15](https://github.com/wilfred-dev/wilfred/issues/15) Added ability to send a single command to STDIN of server without attaching to server console.
 
 ## v0.1.1 (released on 2020-01-11)