+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Please refer to the [official documentation](https://wilfred.readthedocs.io/en/latest/development/) for more information about the CHANGELOG and releases.
 
-## next-release
+## v0.2.0 (released on 2020-01-18)
 
 ### Added